requirements.txt
matplotlib
opencv-python
pandas
numpy
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
pillow
tqdm
